Restructure components into sub-folders
