+++ src/App.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Header from './components/Header'
+import Header from './components/header/Header'
 import TransactionOverviewPage from './components/pages/TransactionOverviewPage'
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom'
 import AddTransactionPage from './components/pages/AddTransactionPage'
