+++ src/components/pages/AddTransactionPage.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react'
-import TransactionFormInput from '../TransactionFormInput'
-import ToggleSwitch from '../ToggleSwitch'
+import TransactionFormInput from '../form/TransactionFormInput'
+import ToggleSwitch from '../toggleSwitch/ToggleSwitch'
 import ReturnToOverviewButton from '../buttons/ReturnToOverviewButton'
 import styled from 'styled-components/macro'
 import { postNewTransactionEntry } from '../utils/services'
@@ -15,9 +15,11 @@ export default function AddTransactionPage() {
       <ReturnToOverviewButton />
     </PageStyled>
   )
+
   function handleToggle() {
     setSelected(!selected)
   }
+
   function onSaveAddTransactionEntry(newTransactionValue, notes, tag) {
     const newTransaction = {
       type: selected ? 'income' : 'spending',
