+++ src/components/pages/EditTransactionPage.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react'
-import TransactionFormInput from '../TransactionFormInput'
-import ToggleSwitch from '../ToggleSwitch'
+import TransactionFormInput from '../form/TransactionFormInput'
+import ToggleSwitch from '../toggleSwitch/ToggleSwitch'
 import ReturnToOverviewButton from '../buttons/ReturnToOverviewButton'
 import styled from 'styled-components/macro'
 import { editTransactionEntry, getTransactionEntry } from '../utils/services'
