+++ src/components/transactionList/TransactionList.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import Transaction from './Transaction'
+import Transaction from '../singleTransaction/Transaction'
 import styled from 'styled-components/macro'
 import PropTypes from 'prop-types'
